simplify config.yaml
