+++ .npmignore
@@ -2,4 +2,4 @@
 .github
 .npmrc
 .git
-doc
+design
