+++ README.md
@@ -1,1 +1,6 @@
 # TODO
+
+## Objetivo
+
+Generar consultas a la database a partir de expressiones lambda de la misma forma que lo hariamos con un array de objetos,
+de esta forma no requiere aprender un nuevo lenguaje de consultas y podremos aprovechar el intelicense.
