+++ design/research.md
