+++ src/orm/model/config.ts
@@ -2,8 +2,6 @@ export interface Database
 {
 	name: string
 	dialect: string
-	disable?:boolean
-	connectionSource?:'env'|'direct'
 	connection:any
 	schema:string
 }
