+++ src/test/task/generateDataForTest.ts
@@ -1030,7 +1030,6 @@ export async function apply(configPath: string, databases: string[], callback: a
   // //string functions
   // //datetime functions
   // //nullables functions 
-
   // OLDS
   // await modify(orm)
   // await crud(orm)
